Added EntityAppearanceColor enum and refactored
