+++ MojioSDK/Models/Pets/Asset.swift
@@ -22,70 +22,6 @@ public enum AssetType: String, Codable {
     case other = "Other"
 }
 
-public protocol DisplayDetailsModel: Codable {
-    var showOnMap: Bool? { get }
-    var color: String? { get }
-    var icon: String? { get }
-    var profileImage: String? { get }
-}
-
-public struct DisplayDetails: DisplayDetailsModel {
-    public let showOnMap: Bool?
-    public let color: String?
-    public let icon: String?
-    public let profileImage: String?
-    
-    public enum CodingKeys: String, CodingKey {
-        case showOnMap = "ShowOnMap"
-        case color = "Color"
-        case icon = "Icon"
-        case profileImage = "ProfileImage"
-    }
-}
-
-public struct DisplayDetailsUpdate: Codable {
-    public var showOnMap: Bool?
-    public var color: String?
-    public var icon: String?
-    public var profileImage: String?
-    
-    public enum CodingKeys: String, CodingKey {
-        case showOnMap = "ShowOnMap"
-        case color = "Color"
-        case icon = "Icon"
-        case profileImage = "ProfileImage"
-    }
-    
-    public init(from displayDetailsModel: DisplayDetailsModel? = nil) {
-        self.init(
-            showOnMap: displayDetailsModel?.showOnMap,
-            color: displayDetailsModel?.color,
-            icon: displayDetailsModel?.icon,
-            profileImage: displayDetailsModel?.profileImage
-        )
-    }
-    
-    public init(
-        showOnMap: Bool? = nil,
-        color: String? = nil,
-        icon: String? = nil,
-        profileImage: String? = nil) {
-        
-        self.showOnMap = showOnMap
-        self.color = color
-        self.icon = icon
-        self.profileImage = profileImage
-    }
-    
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        
-        try container.encodeIfPresent(self.showOnMap, forKey: .showOnMap)
-        try container.encodeIfPresent(self.color, forKey: .color)
-        try container.encodeIfPresent(self.icon, forKey: .icon)
-        try container.encodeIfPresent(self.profileImage, forKey: .profileImage)
-    }
-}
 public protocol AssetModel: Codable, PrimaryKey {
     
     associatedtype L: PetsLocationModel
